Improve methods and documentation.
More options for running co-processor lists. Finish can not wait for the co-processor to complete, begin function added, swap function can optionally call finish as well.
All commands to the co-processor are cached to be written in a burst, this can slightly speed up operation as pyftdi is not toggling chip select when this occurs.
Mew parameters to apprunnner to specify the connector and to set the width and height of the display (it won't work with all displays yet as reg_hcycle and reg_vcycle will need tweaked).