Add in API command ram_cmd to send arbitrary data to RAM_CMD via REG_CMDB_WRITE
This simplifies the FT4222 connector and adds a specific method for sending any size of data buffer to the co-processor. It will wait for the co-processor to process data and feed the RAM_CMD area. Added option of inc for increment address pointer in the connector wr command. REG_CMDB_WRITE does not increment but other areas (RAM_G) will do.