Use stated EVE frequency for sleepclocks. Tweaks for FT4222H speed and stability in QuadSPI mode
FT4222H QuadSPI changes not re-tested on MPSSE or D2XX at present. Change is slightly different approach to chip-select to allow for getspace to be called from within the write() function in eve.py. If CS was selected by the calling function then we'd need to deassert CS then reassert it when calling getspace().